Moving scripts to a specific folder
